create multiple nii files
Create multiple nii files if data is too big for memory. Each saved file should have the starting frame appended in the name. No new input needed to run this (replace tiff_to_nii.py) in code. Should work the same as tiff_to_nii.py if the number of frames is <3384. Only change will be in the nii save name which will have a 0 at the end.